# docs/generated/openapi.rapidapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Naruto Mythos TCG Developer API",
    "description": "A developer-friendly REST API for querying the Naruto Mythos TCG catalogue. Card is a canonical identity; Printing is a collectible identity; Edition is scoped to a Set.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/sets/{public_id}/cards": {
      "get": {
        "tags": [
          "Sets"
        ],
        "summary": "List a set's cards",
        "description": "Paginated list of cards belonging to one set. Supports sorting by number, name, or rarity. card_number is sorted lexically (as a string), never as a number.",
        "operationId": "list_set_cards_v1_sets__public_id__cards_get",
        "parameters": [
          {
            "name": "public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Page number, starting at 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, starting at 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Items per page, up to 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Items per page, up to 100."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field: number, name, or rarity.",
              "default": "number",
              "title": "Sort"
            },
            "description": "Sort field: number, name, or rarity."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort order: asc or desc.",
              "default": "asc",
              "title": "Order"
            },
            "description": "Sort order: asc or desc."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCardsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or sort parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "pagination": {
                    "value": {
                      "error": {
                        "code": "INVALID_PAGINATION",
                        "message": "page must be >= 1."
                      }
                    }
                  },
                  "sort": {
                    "value": {
                      "error": {
                        "code": "INVALID_SORT",
                        "message": "Invalid sort field."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Set not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SET_NOT_FOUND",
                    "message": "Set not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cards": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "List cards",
        "operationId": "list_cards_v1_cards_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Page number, starting at 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, starting at 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Items per page: 1 through 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Items per page: 1 through 100."
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Case-insensitive partial name match.",
              "title": "Name"
            },
            "description": "Case-insensitive partial name match."
          },
          {
            "name": "set",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Exact public set ID, case-insensitive.",
              "title": "Set"
            },
            "description": "Exact public set ID, case-insensitive."
          },
          {
            "name": "number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Exact card number (a string).",
              "title": "Number"
            },
            "description": "Exact card number (a string)."
          },
          {
            "name": "rarity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Case-insensitive Printing rarity/class filter; results remain unique canonical Cards.",
              "title": "Rarity"
            },
            "description": "Case-insensitive Printing rarity/class filter; results remain unique canonical Cards."
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Case-insensitive exact card type.",
              "title": "Type"
            },
            "description": "Case-insensitive exact card type."
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Case-insensitive exact keyword slug.",
              "title": "Keyword"
            },
            "description": "Case-insensitive exact keyword slug."
          },
          {
            "name": "variant",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Case-insensitive Printing treatment/type; at least one Printing must match.",
              "title": "Variant"
            },
            "description": "Case-insensitive Printing treatment/type; at least one Printing must match."
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Case-insensitive exact parent set language, not variant language.",
              "title": "Language"
            },
            "description": "Case-insensitive exact parent set language, not variant language."
          },
          {
            "name": "edition",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Edition public ID; returns unique canonical Cards having a Printing in that Edition.",
              "title": "Edition"
            },
            "description": "Edition public ID; returns unique canonical Cards having a Printing in that Edition."
          },
          {
            "name": "chakra_min",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "description": "Inclusive minimum chakra, >= 0; excludes null chakra.",
              "title": "Chakra Min"
            },
            "description": "Inclusive minimum chakra, >= 0; excludes null chakra."
          },
          {
            "name": "chakra_max",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "description": "Inclusive maximum chakra, >= 0; excludes null chakra.",
              "title": "Chakra Max"
            },
            "description": "Inclusive maximum chakra, >= 0; excludes null chakra."
          },
          {
            "name": "power_min",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "description": "Inclusive minimum power, >= 0; excludes null power.",
              "title": "Power Min"
            },
            "description": "Inclusive minimum power, >= 0; excludes null power."
          },
          {
            "name": "power_max",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "description": "Inclusive maximum power, >= 0; excludes null power.",
              "title": "Power Max"
            },
            "description": "Inclusive maximum power, >= 0; excludes null power."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "number, name, rarity, set (name), or release_date.",
              "default": "number",
              "title": "Sort"
            },
            "description": "number, name, rarity, set (name), or release_date."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "asc or desc; null values always last.",
              "default": "asc",
              "title": "Order"
            },
            "description": "asc or desc; null values always last."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCardsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination, sort, or filter parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "pagination": {
                    "value": {
                      "error": {
                        "code": "INVALID_PAGINATION",
                        "message": "page must be >= 1."
                      }
                    }
                  },
                  "sort": {
                    "value": {
                      "error": {
                        "code": "INVALID_SORT",
                        "message": "Invalid sort field."
                      }
                    }
                  },
                  "chakra_range": {
                    "value": {
                      "error": {
                        "code": "INVALID_FILTER",
                        "message": "chakra_min cannot be greater than chakra_max."
                      }
                    }
                  },
                  "power_range": {
                    "value": {
                      "error": {
                        "code": "INVALID_FILTER",
                        "message": "power_min cannot be greater than power_max."
                      }
                    }
                  },
                  "negative_stat": {
                    "value": {
                      "error": {
                        "code": "INVALID_FILTER",
                        "message": "chakra_min must be greater than or equal to 0."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cards/random": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Get a random card",
        "operationId": "random_card_v1_cards_random_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetail"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No cards available.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "CARD_NOT_FOUND",
                    "message": "No cards available."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/cards/{public_id}": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Get a single card",
        "operationId": "get_card_v1_cards__public_id__get",
        "parameters": [
          {
            "name": "public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetail"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Card not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CARD_NOT_FOUND",
                    "message": "Card not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/keywords/{slug}/cards": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "List a keyword's cards",
        "operationId": "keyword_cards_v1_keywords__slug__cards_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Page number, >= 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, >= 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Items per page, 1 through 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Items per page, 1 through 100."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "number, name, rarity, set, release_date.",
              "default": "number",
              "title": "Sort"
            },
            "description": "number, name, rarity, set, release_date."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "asc or desc; nulls last.",
              "default": "asc",
              "title": "Order"
            },
            "description": "asc or desc; nulls last."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCardsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or sort parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "pagination": {
                    "value": {
                      "error": {
                        "code": "INVALID_PAGINATION",
                        "message": "page must be >= 1."
                      }
                    }
                  },
                  "sort": {
                    "value": {
                      "error": {
                        "code": "INVALID_SORT",
                        "message": "Invalid sort field."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Keyword not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "KEYWORD_NOT_FOUND",
                    "message": "Keyword not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CardDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "TEST-001"
          },
          "number": {
            "type": "string",
            "title": "Number",
            "example": "001"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Test Character Alpha"
          },
          "subtitle": {
            "type": "string",
            "nullable": true,
            "title": "Subtitle",
            "example": null
          },
          "type": {
            "type": "string",
            "nullable": true,
            "title": "Type",
            "example": "Character"
          },
          "rarity": {
            "type": "string",
            "nullable": true,
            "title": "Rarity",
            "example": "Rare"
          },
          "chakra": {
            "type": "integer",
            "nullable": true,
            "title": "Chakra",
            "example": 4
          },
          "power": {
            "type": "integer",
            "nullable": true,
            "title": "Power",
            "example": 5
          },
          "points": {
            "type": "integer",
            "nullable": true,
            "title": "Points",
            "example": null
          },
          "faction": {
            "type": "string",
            "nullable": true,
            "title": "Faction",
            "example": "Test Village"
          },
          "ability_text": {
            "type": "string",
            "nullable": true,
            "title": "Ability Text",
            "example": null
          },
          "flavor_text": {
            "type": "string",
            "nullable": true,
            "title": "Flavor Text",
            "example": null
          },
          "artist": {
            "type": "string",
            "nullable": true,
            "title": "Artist",
            "example": null
          },
          "set": {
            "$ref": "#/components/schemas/SetSummary"
          },
          "keywords": {
            "items": {
              "$ref": "#/components/schemas/KeywordResponse"
            },
            "type": "array",
            "title": "Keywords"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/CardVariantResponse"
            },
            "type": "array",
            "title": "Variants"
          },
          "printings": {
            "items": {
              "$ref": "#/components/schemas/PrintingResponse"
            },
            "type": "array",
            "title": "Printings",
            "description": "Collectible Printings for this canonical Card; `variants` is retained as a legacy alias."
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CardImageResponse"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "name",
          "set"
        ],
        "title": "CardDetail",
        "example": {
          "chakra": 4,
          "faction": "Test Village",
          "id": "TEST-001",
          "images": [],
          "keywords": [
            {
              "name": "Test Keyword",
              "slug": "test-keyword"
            }
          ],
          "name": "Test Character Alpha",
          "number": "001",
          "power": 5,
          "printings": [],
          "rarity": "Rare",
          "set": {
            "code": "TST1",
            "edition": "1st Edition",
            "id": "test-set-1",
            "name": "Test Set Alpha"
          },
          "type": "Character",
          "variants": []
        }
      },
...
      "CardSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "TEST-001"
          },
          "number": {
            "type": "string",
            "title": "Number",
            "example": "001"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Test Character Alpha"
          },
          "subtitle": {
            "type": "string",
            "nullable": true,
            "title": "Subtitle",
            "example": null
          },
          "type": {
            "type": "string",
            "nullable": true,
            "title": "Type",
            "example": "Character"
          },
          "rarity": {
            "type": "string",
            "nullable": true,
            "title": "Rarity",
            "example": "Rare"
          },
          "set": {
            "$ref": "#/components/schemas/SetSummary"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CardImageResponse"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "name",
          "set"
        ],
        "title": "CardSummary",
        "example": {
          "id": "TEST-001",
          "images": [],
          "name": "Test Character Alpha",
          "number": "001",
          "rarity": "Rare",
          "set": {
            "code": "TST1",
            "edition": "1st Edition",
            "id": "test-set-1",
            "name": "Test Set Alpha"
          },
          "type": "Character"
        }
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeywordResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug",
            "example": "test-keyword"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Test Keyword"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "title": "KeywordResponse",
        "example": {
          "name": "Test Keyword",
          "slug": "test-keyword"
        }
      },
...
      "PaginatedCardsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CardSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "PaginatedCardsResponse",
        "example": {
          "data": [],
          "pagination": {
            "has_next": false,
            "has_previous": false,
            "limit": 50,
            "page": 1,
            "pages": 0,
            "total": 0
          }
        }
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "example": 50
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "example": 630
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "example": 13
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "example": true
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous",
            "example": false
          }
        },
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta",
        "example": {
          "has_next": true,
          "has_previous": false,
          "limit": 50,
          "page": 1,
          "pages": 13,
          "total": 630
        }
      },
...
    }
  }
}